% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_genomic_track.R
\name{import_genomic_track}
\alias{import_genomic_track}
\title{Imports a genomic track file using the import functionality in rtracklayer package.}
\usage{
import_genomic_track(track_file, file_format)
}
\arguments{
\item{track_file}{Character string indicating the filename of input genomic track file}

\item{file_format}{Character string indicating the file format of input file like bed, gff, wig. For all formats, see import function in rtracklayer package.}
}
\value{
A GRanges object of input track file.
}
\description{
\code{import_genomic_track} Imports a genomic track file using the import functionality in rtracklayer package.
}
\examples{
mybedfile <- import_genomic_track('mybedfile.bed',format='bed')
}
