% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_genome_build.R
\name{predict_genome_build}
\alias{predict_genome_build}
\title{Predicts the genome build version of genomic track files, if any of the sequence coordinates in the input track file are unique to a genome build version.}
\usage{
predict_genome_build(input_track, species = c("human", "mouse", "drosophila"))
}
\arguments{
\item{input_track}{An input genomic track in the form of a GRanges object; import_genomic_track function could be used to generate this.}

\item{species}{Character string indicating the species of the input genomic track.}
}
\value{
A character string indicating the genome build version of the genomic track.
}
\description{
\code{predict_genome_build} Predicts the genome build version of genomic track files, if any of the sequence coordinates in the input track file are unique to a genome build version.
}
\examples{
build_version <- predict_genome_build(my_input_track,species='human')
}
