% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_human_build.R
\name{predict_human_build}
\alias{predict_human_build}
\title{Predicts the genome build version of genomic track files for human genome, if any of the sequence coordinates in the input track file are unique to a genome build version.}
\usage{
predict_human_build(input_track)
}
\arguments{
\item{input_track}{An input genomic track in the form of a GRanges object; import_genomic_track function could be used to generate this.}
}
\value{
A character string indicating the genome build version of the genomic track.
}
\description{
\code{predict_human_build} Predicts the human genome build version of genomic track files, if any of the sequence coordinates in the input track file are unique to a genome build version.
}
\examples{
build_version <- predict_human_build(my_input_track)
}
